requirements.txt
absl-py==1.4.0
accelerate==0.25.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
audiomentations==0.32.0
audioread==3.0.0
Babel==2.13.1
backcall==0.2.0
blis==0.7.9
cachetools==5.3.0
calmsize==0.1.3
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.2.0
confection==0.0.4
confidence-intervals==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
Cython==3.0.2
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
evaluate==0.4.0
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.19.0
ffmpeg-python==0.2.0
filelock==3.8.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.19.4
idna==3.4
ipykernel==6.27.1
ipython==8.14.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
julius==0.2.7
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
nnAudio==0.3.2
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
overrides==7.4.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pedalboard==0.8.6
pickleshare==0.7.5
Pillow==9.3.0
plac==1.3.5
platformdirs==4.0.0
pooch==1.6.0
praatio==3.8.0
preshed==3.0.8
primePy==1.3
prometheus-client==0.19.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
pylangacq==0.18.0
pyloudnorm==0.1.1
pyparsing==3.0.9
pyroomacoustics==0.7.3
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-memlab==0.2.4
pytz==2022.7.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.10.0
Send2Trash==1.8.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
terminado==0.18.0
tgt==1.4.4
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.3.1
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
# torch==2.1.2+cu121
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
# torchaudio==2.1.2+cu121
# torchvision==0.16.2+cu121
tornado==6.4
tqdm==4.64.1
traitlets==5.9.0
transformers==4.36.1
typer==0.7.0
types-python-dateutil==2.8.19.14
typing_extensions==4.4.0
uri-template==1.3.0
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.2.2
widgetsnbextension==4.0.9
xxhash==3.2.0
yarl==1.8.2
